Share GPOS table as much as possible for font collections
Spacing tables may differ across fonts in a font collection
(TTC) because each font may have different cmap. But having
different data in GPOS prevents sharing the table, which can
increase the font size. When the original font has shared
GPOS, the increase is not ignorable.

The new logic in this patch creates one `EastAsianSpacing`
for each file offset in the original file, and unites all
glyphs into that so that we add a same set of data to shared
GPOS tables.

`NotoSerifCJK-Regular.ttc` is such an example. Before this
patch, single shared ~54kb GPOS tables was copied to 3 tables
and therefore increased the file size by ~100kb. With this
change, the GPOS table is shared.