Change `skip_monospace_ascii` default to `False`
It turned out that most monospace ASCII fonts are intended for regular
uses, that this patch changes the default value of `skip_monospace_ascii`
to `False`. For example, in Microsoft fonts, "P" suffix means
proportional not only for ASCII but also for CJK characters that it is
quite possible to use monospace ASCII variation for regular body text.

Noto Mono is different from regular Noto only for ASCII, and that the
default `Config` still set to `True` only for Noto.