Keep `GlyphData` in `GlyphSets`
This patch changes `GlyphSets` to keep all `GlyphData`.

Before this change, it only keeps glyph ID. It was for efficiencies,
and was enough. However, not keeping advances and some other information
of glyphs became a limitation when adding new capabilities. Issue #38
for example.

The `GlyphDataList` was added to keep all `GlyphData`, while allowing
`set`-like operations.